Added Authentication And Authorization

+++ src/main/java/com/yash/projects/airBnb/service/HotelServiceImplementation.java
@@ -5,13 +5,16 @@ import com.yash.projects.airBnb.dto.HotelInfoDto;
 import com.yash.projects.airBnb.dto.RoomDTO;
 import com.yash.projects.airBnb.entity.Hotel;
 import com.yash.projects.airBnb.entity.Room;
+import com.yash.projects.airBnb.entity.User;
 import com.yash.projects.airBnb.exception.ResourceNotFoundException;
+import com.yash.projects.airBnb.exception.UnAuthorizedException;
 import com.yash.projects.airBnb.repository.HotelRepository;
 import com.yash.projects.airBnb.repository.RoomRepository;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.modelmapper.ModelMapper;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -30,6 +33,8 @@ public class HotelServiceImplementation implements HotelService {
         log.info("Creating new hotel with name: {}", hotelDto.getName());
         Hotel hotel = modelMapper.map(hotelDto,Hotel.class);
         hotel.setActive(false);
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        hotel.setOwner(user);
         hotel=hotelRepository.save(hotel);
         log.info("Hotel created with ID: {}", hotel.getId());
         return modelMapper.map(hotel,HotelDTO.class);
@@ -40,6 +45,11 @@ public class HotelServiceImplementation implements HotelService {
         log.info("Getting hotel with ID: {}", id);
         Hotel hotel = hotelRepository.findById(id)
                 .orElseThrow(()->new ResourceNotFoundException("Hotel not found with ID: " + id));
+
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(hotel.getOwner())){
+            throw new UnAuthorizedException("User is not owner of this hotel");
+        }
         return modelMapper.map(hotel,HotelDTO.class);
     }
 
@@ -49,6 +59,11 @@ public class HotelServiceImplementation implements HotelService {
         Hotel existingHotel = hotelRepository
                 .findById(id)
                 .orElseThrow(()->new ResourceNotFoundException("Hotel not found with ID: " + id));
+
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(existingHotel.getOwner())){
+            throw new UnAuthorizedException("User is not owner of this hotel");
+        }
 
         modelMapper.map(hotel,existingHotel);
         existingHotel.setId(id);
@@ -67,6 +82,11 @@ public class HotelServiceImplementation implements HotelService {
         Hotel hotel = hotelRepository.findById(id)
                 .orElseThrow(()->new ResourceNotFoundException("Hotel not found with ID: " + id));
 
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(hotel.getOwner())){
+            throw new UnAuthorizedException("User is not owner of this hotel");
+        }
+
         for(Room rooms : hotel.getRooms() ){
             inventoryService.deleteAllInventories(rooms);
             roomRepository.deleteById(rooms.getId());
@@ -83,6 +103,11 @@ public class HotelServiceImplementation implements HotelService {
         log.info("Activating hotel with ID: {}", id);
         Hotel hotel = hotelRepository.findById(id)
                 .orElseThrow(()->new ResourceNotFoundException("Hotel not found with ID: " + id));
+
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(hotel.getOwner())){
+            throw new UnAuthorizedException("User is not owner of this hotel");
+        }
         hotel.setActive(true);
 
         //assuming only do it once
