+++ src/main/java/com/yash/projects/airBnb/service/RoomServiceImplementation.java
@@ -4,12 +4,15 @@ package com.yash.projects.airBnb.service;
 import com.yash.projects.airBnb.dto.RoomDTO;
 import com.yash.projects.airBnb.entity.Hotel;
 import com.yash.projects.airBnb.entity.Room;
+import com.yash.projects.airBnb.entity.User;
 import com.yash.projects.airBnb.exception.ResourceNotFoundException;
+import com.yash.projects.airBnb.exception.UnAuthorizedException;
 import com.yash.projects.airBnb.repository.HotelRepository;
 import com.yash.projects.airBnb.repository.RoomRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.modelmapper.ModelMapper;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -32,6 +35,11 @@ public class RoomServiceImplementation implements RoomService{
         Hotel hotel = hotelRepository
                 .findById(hotelId)
                 .orElseThrow(() -> new ResourceNotFoundException("Hotel not found with ID: "+hotelId));
+
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(hotel.getOwner())){
+            throw new UnAuthorizedException("This user does not owns this hotel" + hotelId);
+        }
         Room room = modelMapper.map(roomDto, Room.class);
         room.setHotel(hotel);
         room = roomRepository.save(room);
@@ -50,6 +58,11 @@ public class RoomServiceImplementation implements RoomService{
                 .findById(hotelId)
                 .orElseThrow(() -> new ResourceNotFoundException("Hotel not found with ID: "+hotelId));
 
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(hotel.getOwner())){
+            throw new UnAuthorizedException("User is not owner of this hotel");
+        }
+
         return hotel.getRooms()
                 .stream()
                 .map((element) -> modelMapper.map(element, RoomDTO.class))
@@ -62,6 +75,7 @@ public class RoomServiceImplementation implements RoomService{
         Room room = roomRepository
                 .findById(roomId)
                 .orElseThrow(() -> new ResourceNotFoundException("Room not found with ID: "+roomId));
+
         return modelMapper.map(room, RoomDTO.class);
     }
 
@@ -72,6 +86,10 @@ public class RoomServiceImplementation implements RoomService{
         Room room = roomRepository
                 .findById(roomId)
                 .orElseThrow(() -> new ResourceNotFoundException("Room not found with ID: "+roomId));
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(room.getHotel().getOwner())){
+            throw new UnAuthorizedException("This user does not own this room with id: "+roomId);
+        }
         inventoryService.deleteAllInventories(room);
         roomRepository.deleteById(roomId);
     }
